Reorganise javascript files
We don't need to create bundles for each js file we have. We only need
1 application js pack.
