+++ app/webpacker/packs/application.js
@@ -3,13 +3,13 @@ require.context('govuk-frontend/govuk/assets');
 import "accessible-autocomplete/dist/accessible-autocomplete.min.css";
 
 import { initAll } from 'govuk-frontend';
-import initResponsibleBodiesAutocomplete from "./responsible-bodies-autocomplete";
+import initResponsibleBodiesAutocomplete from "../scripts/responsible-bodies-autocomplete";
 import '../styles/application.scss';
-import initWarnOnUnsavedChanges from "./warn-on-unsaved-changes";
+import initWarnOnUnsavedChanges from "../scripts/warn-on-unsaved-changes";
 
 initAll();
 initWarnOnUnsavedChanges();
-import {initSelectAllNone} from './mno/extra-mobile-data-requests'
+import {initSelectAllNone} from '../scripts/mno/extra-mobile-data-requests'
 
 initSelectAllNone();
 initResponsibleBodiesAutocomplete();
